Separate tests into tests/error.rs
